Create and share core `button` component
