+++ src/components/buy-crypto.tsx
@@ -1,7 +1,4 @@
-import { css, cx } from '@emotion/css';
-import { AuthButtonProps } from 'components/auth-button';
-import { FractalFLogo } from 'components/fractal-f-logo';
-import { getDefaultButtonStyles } from 'components/styles';
+import { Button, ButtonProps } from 'components/button';
 import { useOnramp } from 'hooks';
 import { FC, HTMLAttributes } from 'react';
 
@@ -16,77 +13,15 @@ interface Props {
    *
    * Possible values: 'light' | 'dark'. Defaults to 'light'.
    */
-  variant?: AuthButtonProps['variant'];
+  theme?: ButtonProps['theme'];
 }
 
-export const BuyCrypto: FC<Props> = ({ buttonProps = {}, variant }) => {
-  const {
-    buttonBackground,
-    buttonBackgroundOnHover,
-    logoBackground,
-    logoFill,
-    textColor,
-  } = getColorsFromVariant(variant);
-
-  const defaultButtonStyles = getDefaultButtonStyles({
-    buttonBackground,
-    buttonBackgroundOnHover,
-    textColor,
-  });
-
+export const BuyCrypto: FC<Props> = ({ buttonProps = {}, theme = 'light' }) => {
   const { openOnrampWindow } = useOnramp();
 
   return (
-    <button
-      {...buttonProps}
-      onClick={openOnrampWindow}
-      className={cx(defaultButtonStyles, buttonProps.className)}
-    >
-      <div
-        className={css`
-          height: 2.25rem;
-          aspect-ratio: 1/1;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          background: ${logoBackground};
-          border-radius: 0.125rem;
-        `}
-      >
-        <FractalFLogo fill={logoFill}></FractalFLogo>
-      </div>
-      <div
-        className={css`
-          padding: 0 0.75rem;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          white-space: nowrap;
-        `}
-      >
-        Buy Crypto
-      </div>
-    </button>
+    <Button {...buttonProps} theme={theme} onClick={openOnrampWindow}>
+      Buy Crypto
+    </Button>
   );
 };
-
-function getColorsFromVariant(variant: AuthButtonProps['variant'] = 'dark') {
-  switch (variant) {
-    case 'light':
-      return {
-        buttonBackground: '#c31a88',
-        buttonBackgroundOnHover: '#d71d96',
-        logoBackground: '#f2059f',
-        logoFill: '#fff',
-        textColor: '#fff',
-      };
-    case 'dark':
-      return {
-        buttonBackground: '#000',
-        buttonBackgroundOnHover: '#181818',
-        logoBackground: '#232323',
-        logoFill: '#f2059f',
-        textColor: '#fff',
-      };
-  }
-}
