+++ src/hooks/public/use-onramp.ts
@@ -3,7 +3,7 @@ import { FractalSDKContext } from 'context/fractal-sdk-context';
 import { useCallback, useContext } from 'react';
 
 const ONRAMP_URL = 'https://fractal.is/onramp';
-const POPUP_HEIGHT_PX = 800;
+const POPUP_HEIGHT_PX = 600;
 
 export const useOnramp = () => {
   const { clientId, user } = useContext(FractalSDKContext);
